third_party/WebKit/LayoutTests/fast/css/computed-width-without-renderer-expected.txt

PASS window.width1 is window.width2
PASS
